Moved MySqlMatchSearchMode to separate file. EF.Functions.Match now states that it can only be executed on relational database and throws exception on client-evaluation. Changed MatchMethodInfo discovery mechanism. Updated using statements according to changed enum location.
