+++ src/EFCore.MySql/Extensions/MySqlDbFunctionsEnums.cs
@@ -1,0 +1,10 @@
+﻿namespace Microsoft.EntityFrameworkCore
+{
+    public enum MySqlMatchSearchMode
+    {
+        WithQueryExpansion,
+        InNaturalLanguageMode,
+        InNaturalLanguageModeWithQueryExpansion,
+        InBooleanMode
+    }
+}
